Fixed text rendering

+++ src/main/scala/com/siigna/web/evaluating/Evaluator.scala
@@ -45,9 +45,9 @@ object Evaluator {
         )
 
       case TextExpr(centerX, centerY, height, text) =>
-        getValue[Int](centerX, env, printer).right.flatMap(x =>
-          getValue[Int](centerY, env, printer).right.flatMap(y =>
-            getValue[Int](height, env, printer).right.flatMap(heightValue =>
+        getValue[Double](centerX, env, printer).right.flatMap(x =>
+          getValue[Double](centerY, env, printer).right.flatMap(y =>
+            getValue[Double](height, env, printer).right.flatMap(heightValue =>
               getValue[Any](text, env, printer).right.flatMap(textValue => {
                 printer.text(x,y,heightValue,textValue)
                 Right(env -> Unit)
